Refactor : "Common쪽 폴더링 구조개선" Common의 BaseTimeEntity 등 폴더링이 안되어 있어 폴더생성 후 위치 수정.
