--- src/main/java/com/TTT/TTT/Common/DelYN.java
@@ -1,5 +1,0 @@
-package com.TTT.TTT.Common;
-
-public enum DelYN {
-    Y,N
-}
