+++ src/main/java/com/TTT/TTT/Common/domain/DelYN.java
@@ -1,0 +1,5 @@
+package com.TTT.TTT.Common.domain;
+
+public enum DelYN {
+    Y,N
+}
