+++ src/main/java/com/TTT/TTT/Common/dtos/CommonDto.java
@@ -1,4 +1,4 @@
-package com.TTT.TTT.Common;
+package com.TTT.TTT.Common.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class CommonErrorDto {
+public class CommonDto {
     private int status_code;  //상태코드
     private String status_message;  //상태메시지
+    private Object result;  //데이터
+
 }
