+++ src/main/java/com/TTT/TTT/ListTap/projectList/controller/PrimaryFeatureController.java
@@ -1,13 +1,5 @@
 package com.TTT.TTT.ListTap.projectList.controller;
 
-import com.TTT.TTT.Common.CommonDto;
-import com.TTT.TTT.ListTap.projectList.domain.PrimaryFeature;
-import com.TTT.TTT.ListTap.projectList.dtos.PrimaryFeatureSaveReq;
-import com.TTT.TTT.ListTap.projectList.service.PrimaryFeatureService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
