+++ src/main/java/com/TTT/TTT/Post/domain/Post.java
@@ -1,7 +1,7 @@
 package com.TTT.TTT.Post.domain;
 
-import com.TTT.TTT.Common.BaseTimeEntity;
-import com.TTT.TTT.Common.DelYN;
+import com.TTT.TTT.Common.domain.BaseTimeEntity;
+import com.TTT.TTT.Common.domain.DelYN;
 import com.TTT.TTT.User.domain.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
