+++ src/main/java/com/TTT/TTT/Post/dtos/PostSaveReq.java
@@ -1,10 +1,9 @@
 package com.TTT.TTT.Post.dtos;
 
-import com.TTT.TTT.Common.DelYN;
+import com.TTT.TTT.Common.domain.DelYN;
 import com.TTT.TTT.Post.domain.Post;
 import com.TTT.TTT.User.domain.User;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
