+++ src/main/java/com/TTT/TTT/Post/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.TTT.TTT.Post.repository;
 
-import com.TTT.TTT.Common.DelYN;
+import com.TTT.TTT.Common.domain.DelYN;
 import com.TTT.TTT.Post.domain.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
