+++ src/main/java/com/TTT/TTT/User/UserController/UserController.java
@@ -1,7 +1,7 @@
 
 package com.TTT.TTT.User.UserController;
 
-import com.TTT.TTT.Common.CommonDto;
+import com.TTT.TTT.Common.dtos.CommonDto;
 import com.TTT.TTT.Common.auth.JwtTokenProvider;
 import com.TTT.TTT.Post.domain.Post;
 import com.TTT.TTT.User.UserService.UserService;
@@ -19,7 +19,6 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
