+++ src/main/java/com/TTT/TTT/User/dtos/UserDetailDto.java
@@ -1,6 +1,6 @@
 package com.TTT.TTT.User.dtos;
 
-import com.TTT.TTT.Common.DelYN;
+import com.TTT.TTT.Common.domain.DelYN;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
